server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
langchain-openai==0.0.2
langchain==0.0.292
requests==2.31.0
python-multipart==0.0.6
pillow==10.0.0
